@@ -16,39 +16,41 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/common": "^4.0.0",
-    "@angular/compiler": "^4.0.0",
-    "@angular/core": "^4.0.0",
-    "@angular/forms": "^4.0.0",
-    "@angular/http": "^4.0.0",
-    "@angular/platform-browser": "^4.0.0",
-    "@angular/platform-browser-dynamic": "^4.0.0",
-    "@angular/router": "^4.0.0",
-    "bootstrap": "^3.3.7",
+    "@angular/common": "^5.2.10",
+    "@angular/compiler": "^5.2.10",
+    "@angular/core": "^5.2.10",
+    "@angular/forms": "^5.2.10",
+    "@angular/http": "^5.2.10",
+    "@angular/platform-browser": "^5.2.10",
+    "@angular/platform-browser-dynamic": "^5.2.10",
+    "@angular/router": "^5.2.10",
+    "bootstrap": "^4.1.1",
     "core-js": "^2.4.1",
     "font-awesome": "^4.7.0",
-    "rxjs": "^5.1.0",
+    "ngx-bootstrap": "^2.0.4",
+    "rxjs": "^6.0.0",
+    "rxjs-compat": "6",
     "zone.js": "^0.8.4"
   },
   "devDependencies": {
     "@angular/cli": "^1.7.4",
-    "@angular/compiler-cli": "^4.0.0",
-    "@types/jasmine": "2.5.38",
-    "@types/node": "~6.0.60",
-    "codelyzer": "4.2.1",
+    "@angular/compiler-cli": "^5.2.10",
+    "@types/jasmine": "2.8.6",
+    "@types/node": "~8.10.11",
+    "codelyzer": "4.3.0",
     "htmllint-cli": "^0.0.7",
-    "jasmine-core": "~2.5.2",
-    "jasmine-spec-reporter": "~3.2.0",
-    "karma": "~1.4.1",
-    "karma-chrome-launcher": "~2.0.0",
+    "jasmine-core": "~3.1.0",
+    "jasmine-spec-reporter": "~4.2.1",
+    "karma": "~2.0.2",
+    "karma-chrome-launcher": "~2.2.0",
     "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "^0.2.0",
+    "karma-coverage-istanbul-reporter": "^1.4.2",
     "karma-jasmine": "~1.1.0",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.1.0",
+    "karma-jasmine-html-reporter": "^1.0.0",
+    "protractor": "~5.3.1",
     "stylelint": "^9.2.0",
     "stylelint-config-recommended": "^2.1.0",
-    "ts-node": "~2.0.0",
+    "ts-node": "~6.0.2",
     "tslint": "5.9.1",
     "tslint-config-airbnb": "^5.8.0",
     "typescript": "~2.8.1"
